refactor: align Code Text entry module
