+++ packages/sdk-components-react/package.json
@@ -24,9 +24,9 @@
       "import": "./lib/components.js"
     },
     "./code-text": {
-      "webstudio": "./src/code-text-highlight.tsx",
-      "types": "./lib/types/code-text-highlight.d.ts",
-      "import": "./lib/code-text-highlight.js"
+      "webstudio": "./src/code-text.tsx",
+      "types": "./lib/types/code-text.d.ts",
+      "import": "./lib/code-text.js"
     },
     "./metas": {
       "webstudio": "./src/metas.ts",
@@ -51,7 +51,7 @@
   },
   "scripts": {
     "build": "vite build --config ../../vite.sdk-components.config.ts",
-    "build:args": "NODE_OPTIONS=--conditions=webstudio generate-arg-types './src/*.tsx !./src/*.stories.tsx !./src/*.test.{ts,tsx} !./src/*.ws.ts !./src/*.ws.ts !./src/*.template.tsx' && oxfmt \"**/*.props.ts\"",
+    "build:args": "NODE_OPTIONS=--conditions=webstudio generate-arg-types './src/*.tsx !./src/*.stories.tsx !./src/*.test.{ts,tsx} !./src/*.ws.ts !./src/*.template.tsx !./src/code-text.tsx' && oxfmt \"**/*.props.ts\"",
     "build:stories": "tsx --conditions=webstudio src/generate-stories.ts && oxfmt \"src/__generated__/*.stories.tsx\"",
     "dts": "tsc --project tsconfig.dts.json",
     "test": "vitest run",
